requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
xgboost>=2.0
lightgbm>=4.0
imbalanced-learn>=0.11
mlflow>=2.10
scikit-optimize>=0.9
jupyter>=1.0
joblib>=1.3
scipy>=1.11
shap>=0.44
openpyxl>=3.1
missingno>=0.5
statsmodels>=0.14
umap-learn>=0.5
lime>=0.2
